+++ lib/garner/cache/context.rb
@@ -2,6 +2,14 @@
 Garner.config.option(:context_key_strategies, {
   :default => [Garner::Strategies::Context::Key::Caller]
 })
+Garner.config.option(:rack_context_key_strategies, {
+  :default => [
+    Garner::Strategies::Context::Key::Caller,
+    Garner::Strategies::Context::Key::RequestGet,
+    Garner::Strategies::Context::Key::RequestPost,
+    Garner::Strategies::Context::Key::RequestPath
+  ]
+})
 
 module Garner
   module Cache