+++ lib/garner/mixins/rack.rb
@@ -1,15 +1,5 @@
 require "garner"
 
-# Set up Garner configuration parameters
-Garner.config.option(:rack_context_key_strategies, {
-  :default => [
-    Garner::Strategies::Context::Key::Caller,
-    Garner::Strategies::Context::Key::RequestGet,
-    Garner::Strategies::Context::Key::RequestPost,
-    Garner::Strategies::Context::Key::RequestPath
-  ]
-})
-
 module Garner
   module Mixins
     module Rack